refactor: Reservations 도메인 객체 삭제
